feat: Add get tag endpoint with pagination and order

+++ src/StackOverflow.TagManagement.Api/StackOverflow.TagManagement.Api/Models/TagMetaData.cs
@@ -1,9 +1,8 @@
 ﻿using MongoDB.Bson.Serialization.Attributes;
-using System.Numerics;
 
 namespace StackOverflow.TagManagement.Api.Models;
 
-public record TagMetaData(BigInteger TotalTagCount)
+public record TagMetaData(int TotalTagCount)
 {
     [BsonId]
     public string Version { get; } = Constants.METADATA_VERSION;
